+++ app/api/files/jsRoutes.js
@@ -1,6 +1,3 @@
-/*eslint-disable max-statements*/
-
-import { isArray } from 'util';
 import Joi from 'joi';
 import multer from 'multer';
 import fs from 'fs';
@@ -10,12 +7,11 @@ import proxy from 'express-http-proxy';
 import entities from 'api/entities';
 import mailer from 'api/utils/mailer';
 import { search } from 'api/search';
-import { CSVLoader, CSVExporter } from 'api/csv';
-import { generateFileName, temporalFilesPath } from 'api/files/filesystem';
+import { CSVLoader } from 'api/csv';
+import { generateFileName } from 'api/files/filesystem';
 import settings from 'api/settings';
 import { processDocument } from 'api/files/processDocument';
 
-import errorLog from 'api/log/errorLog';
 import configPaths from '../config/paths';
 import { validation, handleError, createError } from '../utils';
 import needsAuthorization from '../auth/authMiddleware';
@@ -167,85 +163,6 @@ export default app => {
         });
 
       res.json('ok');
-    }
-  );
-
-  const parseQueryProperty = (query, property) =>
-    query[property] ? JSON.parse(query[property]) : query[property];
-
-  const generateExportFileName = databaseName => `${databaseName}-${new Date().toISOString()}.csv`;
-
-  const removeTempFile = filePath => () => {
-    fs.unlink(filePath, err => {
-      if (err) errorLog.error(`Error unlinking exported file: ${filePath}`);
-    });
-  };
-
-  app.get(
-    '/api/export',
-    validation.validateRequest({
-      properties: {
-        query: {
-          properties: {
-            filters: { type: 'string' },
-            types: { type: 'string' },
-            _types: { type: 'string' },
-            fields: { type: 'string' },
-            allAggregations: { type: 'string' },
-            userSelectedSorting: { type: 'string' },
-            aggregations: { type: 'string' },
-            order: { type: 'string' },
-            sort: { type: 'string' },
-            limit: { type: 'string' },
-            searchTerm: { type: 'string' },
-            includeUnpublished: { type: 'string' },
-            treatAs: { type: 'string' },
-            unpublished: { type: 'string' },
-            select: { type: 'array', items: [{ type: 'string' }] },
-            ids: { type: 'string' },
-          },
-        },
-      },
-    }),
-    (req, res, next) => {
-      req.query.filters = parseQueryProperty(req.query, 'filters');
-      req.query.types = parseQueryProperty(req.query, 'types');
-      req.query.fields = parseQueryProperty(req.query, 'fields');
-      req.query.aggregations = parseQueryProperty(req.query, 'aggregations');
-      req.query.select = parseQueryProperty(req.query, 'select');
-      req.query.unpublished = parseQueryProperty(req.query, 'unpublished');
-      req.query.includeUnpublished = parseQueryProperty(req.query, 'includeUnpublished');
-
-      req.query.ids = parseQueryProperty(req.query, 'ids');
-      if (!isArray(req.query.ids)) delete req.query.ids;
-
-      Promise.all([search.search(req.query, req.language, req.user), settings.get()]).then(
-        // eslint-disable-next-line camelcase
-        ([results, { dateFormat, site_name }]) => {
-          const exporter = new CSVExporter();
-
-          const temporalFilePath = temporalFilesPath(
-            generateFileName({ originalname: 'export.csv' })
-          );
-          const fileStream = fs.createWriteStream(temporalFilePath, { emitClose: true });
-
-          const exporterOptions = { dateFormat, language: req.language };
-
-          exporter
-            .export(results, req.query.types, fileStream, exporterOptions)
-            .then(() => {
-              res.download(
-                temporalFilePath,
-                generateExportFileName(site_name),
-                removeTempFile(temporalFilePath)
-              );
-            })
-            .catch(e => {
-              removeTempFile(temporalFilePath)();
-              next(e);
-            });
-        }
-      );
     }
   );
 };
