+++ app/api/utils/testingRoutes.ts
@@ -14,13 +14,14 @@ declare global {
 
 const iosocket = jasmine.createSpyObj('socket', ['emit']);
 
-const setUpApp = (route: Function): Application => {
+const setUpApp = (route: Function, ...customMiddleware: ((req: Request, _es: Response, next: NextFunction) =>void)[] ): Application => {
   const app: Application = express();
   app.use(bodyParser.json());
   app.use((req: Request, _res: Response, next: NextFunction) => {
     req.getCurrentSessionSockets = () => ({ sockets: [iosocket], emit: iosocket.emit });
     next();
   });
+  customMiddleware.forEach(customMiddleware => app.use(customMiddleware));
   route(app);
   app.use(errorHandlingMiddleware);
   return app;
